Modify TestConfig to use a separate TestConfigTester app
